+++ passes/S015/testdata/src/a/comment_ignore.go
@@ -1,0 +1,22 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:S015
+	_ = map[string]*schema.Schema{
+		"INVALID": {},
+	}
+
+	//lintignore:S015
+	_ = map[string]*schema.Schema{
+		"invalid!": {},
+	}
+
+	//lintignore:S015
+	_ = map[string]*schema.Schema{
+		"invalid-name": {},
+	}
+}