+++ passes/S015/testdata/src/a/outside_package.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = map[string]*schema.Schema{
+		"INVALID": {},
+	}
+
+	_ = map[string]*schema.Schema{
+		"invalid!": {},
+	}
+
+	_ = map[string]*schema.Schema{
+		"invalid-name": {},
+	}
+}