+++ passes/S015/testdata/src/a/schema/schema.go
@@ -1,0 +1,3 @@
+package schema
+
+type Schema struct{}